requirements.txt
# LLM API
openai>=1.0.0

# Data
datasets>=2.14.0
pandas>=2.0.0

# Utils
python-dotenv>=1.0.0
tqdm>=4.65.0

# Caching
diskcache>=5.6.0
